+++ module/apps/configure-check.js
@@ -1,40 +1,28 @@
-import { getAbilityCheckOptions } from "../game/effect-flags.js";
-import { getDisplayableCheckComponents } from "../game/check-and-save-components.js";
-import { i18n } from "../utils.js";
 
 export class ConfigureCheck extends Dialog {
-    constructor(actor, ability, activation, options) {
+    constructor(config, options) {
         super({}, options);
 
-        this.actor = actor;
-        this.ability = ability;
-        this.activation = activation;
+        this.config = config;
 
-        const checkOptions = getAbilityCheckOptions(actor, ability, activation);
-        const config = activation?.config || {};
-
-        const advantage = !config.check?.disadvantage && (checkOptions.advantage || config.check?.advantage);
-        const disadvantage = !config.check?.advantage && (checkOptions.disadvantage || config.check?.disadvantage);
-
-        const defaultMode = advantage ? "advantage" : (disadvantage ? "disadvantage" : "normal");
+        const defaultMode = config.advantage ? "advantage" : (config.disadvantage ? "disadvantage" : "normal");
 
         this.data = {
             buttons: {
                 advantage: {
                     label: game.i18n.localize("DND5E.Advantage"),
-                    callback: html => this.resolve(this._onSubmitRoll(html, "advantage"))
+                    callback: html => this.resolve(this._onSubmitRoll(html, CONFIG.Dice.D20Roll.ADV_MODE.ADVANTAGE))
                 },
                 normal: {
                     label: game.i18n.localize("DND5E.Normal"),
-                    callback: html => this.resolve(this._onSubmitRoll(html, "normal"))
+                    callback: html => this.resolve(this._onSubmitRoll(html, CONFIG.Dice.D20Roll.ADV_MODE.NORMAL))
                 },
                 disadvantage: {
                     label: game.i18n.localize("DND5E.Disadvantage"),
-                    callback: html => this.resolve(this._onSubmitRoll(html, "disadvantage"))
+                    callback: html => this.resolve(this._onSubmitRoll(html, CONFIG.Dice.D20Roll.ADV_MODE.DISADVANTAGE))
                 }
             },
             default: defaultMode,
-            ...options
         }
     }
 
@@ -42,18 +30,18 @@ export class ConfigureCheck extends Dialog {
         return foundry.utils.mergeObject(super.defaultOptions, {
             template: "modules/wire/templates/apps/configure-check.hbs",
             classes: ["dialog", "configure-check", "configure-ability-check"],
-            width: 300
+            width: 350
         });
     }
     
     get title() {
-        return this.options.title || i18n("wire.configure-check.check-title", { name: CONFIG.DND5E.abilities[this.ability] });
+        return this.options.title || this.config.title;
     }
     
     getData(opts) {
         const { buttons } = super.getData(opts);
 
-        const components = [ ...getDisplayableCheckComponents(this.actor, this.ability) ];
+        const components = this.options.wire.components || [];
 
         return {
             components,
@@ -64,18 +52,11 @@ export class ConfigureCheck extends Dialog {
     _onSubmitRoll(html, mode) {
         const form = html[0].querySelector("form");
 
-        const config = this.activation?.config || {};
-        const bonusInput = form.bonus.value;
+        const bonus = form.bonus.value;
 
-        const advantage = mode === "advantage";
-        const disadvantage = mode === "disadvantage";
-        const normal = mode === "normal";
-        const bonus = [bonusInput, config.check?.bonus].filter(b => b).join(" + ");
+        const attack = { mode, bonus };
 
-        const check = { advantage, disadvantage, normal };
-        if (bonus) { check.parts = [bonus]; }
-
-        this.resolve(foundry.utils.mergeObject(config, check));
+        this.resolve(attack);
         this.close();
         return false;
     }
@@ -89,7 +70,7 @@ export class ConfigureCheck extends Dialog {
     }
 
     async close(options) {
-        this.resolve();
+        this.resolve(null);
         return super.close(options);
     }
 }